requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.87
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
diffrax==0.6.0
docker-pycreds==0.4.0
einops==0.8.0
equinox==0.11.8
etils==1.10.0
filelock==3.16.1
flax==0.10.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
humanize==4.11.0
hydra-core==1.3.2
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources==6.4.5
iniconfig==2.0.0
jax @ file:///home/conda/feedstock_root/build_artifacts/jax_1729336343576/work
jaxlib==0.4.34
jaxtyping==0.2.34
Jinja2==3.1.4
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.8
lineax==0.0.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes @ file:///home/conda/feedstock_root/build_artifacts/ml_dtypes_1726376268746/work
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1727392354687/work
optax==0.2.3
optimistix==0.0.9
orbax-checkpoint==0.8.0
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1729480690820/work/dist/scipy-1.14.1-cp311-cp311-linux_x86_64.whl#sha256=1d6b83e78fb8e82dc54a9de4e73aed118775fa22ccdc582469171f891aa3e912
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tensorstore==0.1.67
toolz==1.0.0
torch==2.5.1
torchmetrics==1.5.1
torchvision==0.20.1
tqdm==4.66.6
triton==3.1.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.5
yarl==1.17.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1726248574750/work
